rename and cleanup
