+++ packages/react-composites/src/composites/CallWithChatComposite/styles/PeoplePaneContent.styles.ts
